requirements.txt
pyTelegramBotAPI
python-dotenv>=1.0.0

